Extract non-runtime helpers from constants
