+++ src/mindroom/__init__.py
@@ -3,7 +3,7 @@
 import os
 from importlib.metadata import version
 
-from mindroom.constants import patch_chromadb_for_python314
+from mindroom.chromadb_compat import patch_chromadb_for_python314
 
 # MindRoom should never emit Agno network telemetry, even if a user .env enables it.
 os.environ["AGNO_TELEMETRY"] = "false"
