+++ src/mindroom/file_ops.py
@@ -1,0 +1,18 @@
+"""Focused filesystem helpers."""
+
+import shutil
+from pathlib import Path
+
+
+def safe_replace(tmp_path: Path, target_path: Path) -> None:
+    """Replace *target_path* with *tmp_path*, with a fallback for bind mounts.
+
+    ``Path.replace`` performs an atomic rename which fails on some filesystems
+    (e.g. Docker bind mounts) with ``OSError: [Errno 16] Device or resource
+    busy``. When that happens we fall back to a non-atomic copy.
+    """
+    try:
+        tmp_path.replace(target_path)
+    except OSError:
+        shutil.copy2(tmp_path, target_path)
+        tmp_path.unlink(missing_ok=True)
